+++ inc/light.h
@@ -6,14 +6,14 @@
 /*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/18 18:39:49 by cle-tron          #+#    #+#             */
-/*   Updated: 2024/12/19 17:30:35 by wlin             ###   ########.fr       */
+/*   Updated: 2024/12/24 16:40:17 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIGHT_H
 # define LIGHT_H
 
-int		color_converter(int color[3], float intensity);
+int		color_converter(int color[3], double intensity);
 void	lightning(t_hit_rec *rec, t_data *data);
 
 #endif
