+++ lib/libft/inc/libft.h
@@ -6,12 +6,12 @@
 /*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 18:27:18 by ale-tron          #+#    #+#             */
-/*   Updated: 2024/12/19 17:33:19 by wlin             ###   ########.fr       */
+/*   Updated: 2024/12/24 17:37:57 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-# include <stdlib.h>
-# define LIBFT_H
+#include <stdlib.h>
+#define LIBFT_H
 
 int				ft_isalpha(int c);
 int				ft_isalnum(int c);
@@ -35,7 +35,7 @@ void			*ft_memchr(const void *s, int c, size_t n);
 int				ft_memcmp(const void *s1, const void *s2, size_t n);
 char			*ft_strnstr(const char *str, const char *to_find, size_t len);
 int				ft_atoi(const char *str);
-float			ft_atod(char *str);
+double			ft_atod(char *str);
 void			*ft_calloc(size_t count, size_t size);
 char			*ft_strdup(const char *s1);
 char			*ft_substr(char const *s, unsigned int start, size_t len);
