+++ src/hit/hit_plane.c
@@ -6,17 +6,17 @@
 /*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/27 16:44:04 by wlin              #+#    #+#             */
-/*   Updated: 2024/12/19 17:30:35 by wlin             ###   ########.fr       */
+/*   Updated: 2024/12/24 17:45:18 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "miniRT.h"
+#include "miniRT.h"
 
-bool hit_plane(t_ray ray, t_obj pl, t_hit_rec *rec)
+bool	hit_plane(t_ray ray, t_obj pl, t_hit_rec *rec)
 {
-	float	alignment;
-	float	plane_t;
-	float	aux;
+	double	alignment;
+	double	plane_t;
+	double	aux;
 
 	alignment = vec_dot(ray.direction, pl.vc);
 	if (fabs(alignment) < 1e-6)
